Changed the parts where u64 was mistakenly used to usize.

+++ crates/soundlog-debuger/src/cui/play.rs
@@ -44,14 +44,14 @@ pub fn play_vgm(
 
     // Register callbacks for all chip types
     callback_stream.on_write(
-        |inst: Instance, spec: chip::PsgSpec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::PsgSpec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!("Sn76489Write({:?}, 0x{:02X})", inst, spec.value);
             print_register_log(sample, &reg_info, event, dry_run);
         },
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Ym2413Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Ym2413Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Ym2413Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -61,7 +61,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Ym2612Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Ym2612Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Ym2612Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -71,7 +71,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Ym2151Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Ym2151Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Ym2151Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -81,7 +81,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Ym2203Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Ym2203Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Ym2203Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -91,7 +91,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Ym2608Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Ym2608Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Ym2608Write({:?}, P0x{:02X}:0x{:02X}=0x{:02X})",
                 inst, spec.port, spec.register, spec.value
@@ -101,7 +101,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Ym2610Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Ym2610Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             // format_command_brief uses Ym2610bWrite for the Ym2610b command; keep the Write style
             let reg_info = format!(
                 "Ym2610bWrite({:?}, P0x{:02X}:0x{:02X}=0x{:02X})",
@@ -112,7 +112,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Ym3812Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Ym3812Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Ym3812Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -122,7 +122,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Ym3526Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Ym3526Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Ym3526Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -132,7 +132,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Y8950Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Y8950Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Y8950Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -142,28 +142,28 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Ymf262Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Ymf262Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!("Ymf262Write({:?}, {:?})", inst, spec);
             print_register_log(sample, &reg_info, event, dry_run);
         },
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Ymf278bSpec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Ymf278bSpec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!("Ymf278bWrite({:?}, {:?})", inst, spec);
             print_register_log(sample, &reg_info, event, dry_run);
         },
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Ymf271Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Ymf271Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!("Ymf271Write({:?}, {:?})", inst, spec);
             print_register_log(sample, &reg_info, event, dry_run);
         },
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Ymz280bSpec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Ymz280bSpec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Ymz280bWrite({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -173,7 +173,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Rf5c68U8Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Rf5c68U8Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Rf5c68U8Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.offset, spec.value
@@ -183,7 +183,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Rf5c68U16Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Rf5c68U16Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Rf5c68U16Write({:?}, 0x{:04X}=0x{:02X})",
                 inst, spec.offset, spec.value
@@ -195,7 +195,7 @@ pub fn play_vgm(
     callback_stream.on_write(
         |inst: Instance,
          spec: chip::Rf5c164U16Spec,
-         sample: u64,
+         sample: usize,
          event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Rf5c164U16Write({:?}, 0x{:04X}=0x{:02X})",
@@ -206,7 +206,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::SegaPcmSpec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::SegaPcmSpec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "SegaPcmWrite({:?}, 0x{:04X}=0x{:02X})",
                 inst, spec.offset, spec.value
@@ -216,7 +216,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::QsoundSpec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::QsoundSpec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "QsoundWrite({:?}, 0x{:04X}=0x{:04X})",
                 inst, spec.register, spec.value
@@ -226,7 +226,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::ScspSpec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::ScspSpec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "ScspWrite({:?}, 0x{:04X}=0x{:04X})",
                 inst, spec.offset, spec.value
@@ -238,7 +238,7 @@ pub fn play_vgm(
     callback_stream.on_write(
         |inst: Instance,
          spec: chip::WonderSwanSpec,
-         sample: u64,
+         sample: usize,
          event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "WonderSwanWrite({:?}, 0x{:04X}=0x{:02X})",
@@ -251,7 +251,7 @@ pub fn play_vgm(
     callback_stream.on_write(
         |inst: Instance,
          spec: chip::WonderSwanRegSpec,
-         sample: u64,
+         sample: usize,
          event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "WonderSwanRegWrite({:?}, 0x{:02X}=0x{:02X})",
@@ -262,7 +262,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::VsuSpec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::VsuSpec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "VsuWrite({:?}, 0x{:04X}=0x{:02X})",
                 inst, spec.offset, spec.value
@@ -272,7 +272,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Saa1099Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Saa1099Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Saa1099Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -282,7 +282,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Es5503Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Es5503Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Es5503Write({:?}, 0x{:04X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -292,7 +292,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Es5506U8Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Es5506U8Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Es5506U8Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -302,7 +302,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Es5506U16Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Es5506U16Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Es5506U16Write({:?}, 0x{:02X}=0x{:04X})",
                 inst, spec.register, spec.value
@@ -312,7 +312,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::X1010Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::X1010Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "X1010Write({:?}, 0x{:04X}=0x{:02X})",
                 inst, spec.offset, spec.value
@@ -322,7 +322,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::C352Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::C352Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "C352Write({:?}, 0x{:04X}=0x{:04X})",
                 inst, spec.register, spec.value
@@ -332,7 +332,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Ga20Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Ga20Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Ga20Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -342,7 +342,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::MultiPcmSpec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::MultiPcmSpec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "MultiPcmWrite({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -352,7 +352,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Upd7759Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Upd7759Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Upd7759Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -362,7 +362,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Okim6258Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Okim6258Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Okim6258Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -372,7 +372,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Okim6295Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Okim6295Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Okim6295Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -382,7 +382,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::K054539Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::K054539Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "K054539Write({:?}, 0x{:04X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -393,7 +393,7 @@ pub fn play_vgm(
 
     // Harmonize write naming with parse output (use XWrite form)
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Huc6280Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Huc6280Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Huc6280Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -403,7 +403,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::C140Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::C140Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "C140Write({:?}, 0x{:04X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -413,7 +413,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::K053260Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::K053260Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "K053260Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -423,7 +423,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::PokeySpec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::PokeySpec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "PokeyWrite({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -433,7 +433,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Ay8910Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Ay8910Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "Ay8910Write({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -443,7 +443,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::GbDmgSpec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::GbDmgSpec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "GbDmgWrite({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -453,7 +453,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::NesApuSpec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::NesApuSpec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "NesApuWrite({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -463,7 +463,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::MikeySpec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::MikeySpec, sample: usize, event: Option<Vec<StateEvent>>| {
             let reg_info = format!(
                 "MikeyWrite({:?}, 0x{:02X}=0x{:02X})",
                 inst, spec.register, spec.value
@@ -473,7 +473,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::Scc1Spec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::Scc1Spec, sample: usize, event: Option<Vec<StateEvent>>| {
             // Keep explicit Scc1Write format used by parse for readability
             let reg_info = format!(
                 "Scc1Write({:?}, P0x{:02X}:0x{:02X}=0x{:02X})",
@@ -484,7 +484,7 @@ pub fn play_vgm(
     );
 
     callback_stream.on_write(
-        |inst: Instance, spec: chip::PwmSpec, sample: u64, event: Option<Vec<StateEvent>>| {
+        |inst: Instance, spec: chip::PwmSpec, sample: usize, event: Option<Vec<StateEvent>>| {
             // Match parse's PwmWrite formatting (show register and 24-bit value)
             let reg_info = format!(
                 "PwmWrite({:?}, reg=0x{:02X}=0x{:06X})",
@@ -522,7 +522,7 @@ pub fn play_vgm(
 }
 
 /// Helper function to print register log line with events
-fn print_register_log(sample: u64, reg_info: &str, events: Option<Vec<StateEvent>>, dry_run: bool) {
+fn print_register_log(sample: usize, reg_info: &str, events: Option<Vec<StateEvent>>, dry_run: bool) {
     let event_str = if let Some(evs) = events {
         if evs.is_empty() {
             String::new()
